+++ apis/server.d.ts
@@ -96,7 +96,7 @@ export function on (event: 'bootstrap', listener: (app: Application) => void): _
 export function once (event: 'bootstrap', listener: (app: Application) => void): _cds
 
 /**
- 	 * Emitted before the model is compiled for usage in Node.js or Java runtime.
+ 	 * Emitted  the model is compiled for usage in Node.js or Java runtime.
       	 * @beta
  	 */
 export function on (event: 'compile.for.runtime', listener: (model: CSN) => void): _cds
@@ -162,9 +162,9 @@ export function exit (): void
 /**
  * @see [capire](https://cap.cloud.sap/docs/node.js/cds-facade#cds-error)
  */
-export function error (code: number, message: string, details?: {stack?: unknown}, caller?: (...args: any[]) => unknown): Error
-export function error (message: string, details?: {code: number, stack?: unknown}, caller?: (...args: any[]) => unknown): Error
-export function error (details: {code: number, message: string}, caller?: (...args: any[]) => unknown): Error
+export function error (status: number, message: string, details?: {stack?: unknown}, caller?: (...args: any[]) => unknown): Error
+export function error (message: string, details?: {status: number, stack?: unknown}, caller?: (...args: any[]) => unknown): Error
+export function error (details: {status: number, message: string}, caller?: (...args: any[]) => unknown): Error
 export function error (string: TemplateStringsArray, ...args: any[]): Error
 
 
@@ -204,6 +204,7 @@ type Middlewares = 'context' | 'trace' | 'auth' | 'ctx_model' | string
 
 export const middlewares: {
   add: (middleware: RequestHandler, pos?: XOR<XOR<{ at: number }, { after: Middlewares }>, { before: Middlewares }>) => void,
+  before: RequestHandler[],
 }
 
 /**