+++ CHANGELOG.md
@@ -9,7 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Added
 ### Changed
 - `ResultHandler` now returns `unknown` instead of `void`, to accommodate asynchronous functions when having `@typescript-eslint/strict-void-return` activated
+- Documentation for `cds.test.axios` mentioning that `@cap-js/cds-test@1` now returns an `axios` facade in absence of `axios`.
+
 ### Deprecated
+- `cds.test.chai`, `cds.test.assert` pointing to either `cds.test.expect` or a custom import of `chai`.
 ### Removed
 ### Fixed
 ### Security